requirements.txt
joblib
ffmpeg-python
openai-whisper
numpy
pandas

# Standard library modules (no need to install)
# json
# os
